backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0

# Environment & Config
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.1

# Document Parsing
PyMuPDF==1.24.5
unstructured[pdf]==0.15.14
python-docx==1.1.2

# LangChain Ecosystem
langchain==0.2.6
langchain-community==0.2.6
langchain-experimental==0.0.62
langchain-openai==0.1.13
langchain-cohere==0.1.9

# LLM Providers
openai==1.35.3
cohere==5.5.8

# Vector DB
qdrant-client==1.9.1

# Sparse Retrieval
rank-bm25==0.2.2

# Evaluation
ragas==0.1.14
datasets>=2.19.0
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
